requirements.txt
aiohttp==3.8.5
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.4
Cython==3.0.2
dl==0.1.0
docutils==0.20.1
Flask==2.3.3
flask_login==0.6.2
flask_minify==0.42
flask_sqlalchemy==3.1.1
HTMLParser==0.0.2
ipython==8.15.0
ipywidgets==8.1.1
Jinja2==3.1.2
jnius==1.1.0
keyring==24.2.0
lockfile==0.12.2
mock==5.1.0
numpy==1.26.0
Pillow==10.0.1
protobuf==4.24.3
pymssql==2.2.8
pyOpenSSL==23.2.0
python-dotenv==1.0.0
PyYAML==6.0.1
railroad==0.5.0
simplejson==3.19.1
Sphinx==7.2.6
SQLAlchemy==2.0.21
tokenize_rt==5.2.0
tornado==6.3.3
trove_classifiers==2023.9.19
truststore==0.8.0
urllib3_secure_extra==0.1.0
uvloop==0.17.0
Werkzeug==2.3.7
xmlrpclib==1.0.1
zstandard==0.21.0
